[DRAFT] Replace includeLocal boolean flag with meaningful constant
The boolean flag is not intuitive. For example, when you
compare HEAD with something else you get two different results
depending on the flag. Contradicting, when you compare "master" with
"stable-1.0" and your currently checked out branch is something else
(eg. "feature-x") then you get the same result whether or not the flag
is true or false.

Also, there is the assumption in Team UI that when you compare workspace
content with something else that workspace is always the "left" /
"source" side of the comparison. The "include local change" flag
lead to wrong internal API usage by specifying HEAD on the "right"
("remote") side.

This commit replaces the boolean flag with a "WORKING_TREE" constant
which can only be used on the left/source side. Additionally, it
introduces an "INDEX" constant for later use to allow comparison with
the content in the Git index.

Change-Id: I36dc1d995fc7e7ddbcc9b15d4f1ff5c71a6025ac
Signed-off-by: Gunnar Wagenknecht <gunnar@wagenknecht.org>